% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.r, R/contributors_from_csv.r
\name{assign_contributors}
\alias{assign_contributors}
\title{Assign contributors to a test}
\usage{
assign_contributors(test_id, contributors, committees)

assign_contributors(test_id, contributors, committees)
}
\arguments{
\item{test_id}{Test ID}

\item{contributors}{List of contributors}

\item{committees}{List of committee names}
}
\value{
List containing test ID and response status

List containing test ID and response status
}
\description{
Assign contributors to a test

Assign contributors to a test
}
