% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.r
\name{assign_contributors_bulk}
\alias{assign_contributors_bulk}
\title{Assign contributors to multiple tests}
\usage{
assign_contributors_bulk(test_ids, contributors)
}
\arguments{
\item{test_ids}{Vector of test IDs}

\item{contributors}{List of contributors}
}
\value{
Data frame with assignment results
}
\description{
Assign contributors to multiple tests
}
