% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_creation_multiple.r
\name{create_multiple_tests}
\alias{create_multiple_tests}
\title{Create multiple tests in Inspera}
\usage{
create_multiple_tests(
  template_id,
  title_prefix,
  start_time,
  duration,
  gap_between_tests
)
}
\arguments{
\item{template_id}{The ID of the template to use for the tests}

\item{title_prefix}{A prefix for the test titles}

\item{start_time}{The start time of the first test (in ISO 8601 format)}

\item{duration}{The duration of each test in minutes}

\item{gap_between_tests}{The time gap between tests in minutes}
}
\value{
A list containing the created test data for each test, or NULL if creation failed
}
\description{
This function creates multiple tests in the Inspera system using the provided parameters.
The number of tests to create is determined by the NUM_TESTS variable in the .env file.
The external_test_id is generated automatically.
}
