% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_creation.r
\name{create_new_test}
\alias{create_new_test}
\title{Create a new test in Inspera}
\usage{
create_new_test(
  template_id,
  external_test_id,
  title,
  start_time,
  end_time,
  duration
)
}
\arguments{
\item{template_id}{The ID of the template to use for the test}

\item{external_test_id}{An external ID for the test}

\item{title}{The title of the test}

\item{start_time}{The start time of the test (in ISO 8601 format)}

\item{end_time}{The end time of the test (in ISO 8601 format)}

\item{duration}{The duration of the test in minutes}
}
\value{
A list containing the created test data, or NULL if creation failed
}
\description{
This function creates a new test in the Inspera system using the provided parameters.
}
