% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{create_test_and_collect_metadata}
\alias{create_test_and_collect_metadata}
\title{Create a new test and collect its metadata}
\usage{
create_test_and_collect_metadata(template_id)
}
\arguments{
\item{template_id}{ID of the template to use}
}
\value{
A list containing the new test information and the path to the metadata file
}
\description{
Create a new test and collect its metadata
}
