% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{.onLoad}
\alias{.onLoad}
\title{Internal function to log package startup messages}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The library name}

\item{pkgname}{The package name}
}
\description{
Internal function to log package startup messages
}
