% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r, R/contributors_from_csv.r, R/utils.r
\name{get_access_token}
\alias{get_access_token}
\title{Get Inspera API access token}
\usage{
get_access_token()

get_access_token()

get_access_token()
}
\value{
Access token as a string

A string containing the new access token

Access token as a string
}
\description{
Get Inspera API access token

Get a new access token

Get Inspera API access token using environment variable
}
