% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_env_var}
\alias{get_env_var}
\title{Get environment variable with a default value}
\usage{
get_env_var(var_name, default_value = NULL)
}
\arguments{
\item{var_name}{Name of the environment variable}

\item{default_value}{Default value if the environment variable is not set}
}
\value{
The value of the environment variable or the default value
}
\description{
Get environment variable with a default value
}
