% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{save_results_csv}
\alias{save_results_csv}
\title{Save results to CSV}
\usage{
save_results_csv(results, file_path)
}
\arguments{
\item{results}{Results to save}

\item{file_path}{File path to save to}
}
\description{
Save results to CSV
}
