% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{validate_env_var}
\alias{validate_env_var}
\title{Validate if a required environment variable is set}
\usage{
validate_env_var(var_name)
}
\arguments{
\item{var_name}{Name of the environment variable}
}
\description{
Validate if a required environment variable is set
}
